Create README.md
This face detector and recognizer is still under improvement by using CNN. Until now, face detector uses cascaded Viola-Jones detector with over 90% of recall and 80% of accuracy. Face recognizer uses PCA combined with KNN, however accuracy of recognition is not decent because the data set is not decent. This is an open-sourced program but the data set cannot be used for commercial use.